test: move tests for integrations into tests/ext (#1097)
## Summary

<!-- Brief description of the changes -->

## Changelog

<!--
User-facing changelog details, rendered below the commit title in CHANGELOG.md.
Delete this section if there are no user-facing changes.
-->

## Test Plan

<!-- How was this tested? -->
