+++ tests/ext/duckdb/test_ducklake.py
