+++ tests/ext/duckdb/test_ducklake_config.py
