+++ tests/ext/duckdb/test_native.py
