+++ tests/ext/lancedb/test_integration.py
